Split tools file to separate files for each class
